get rid of unnecessary controllers, polish error messages
