+++ src/apps/java/consumption/com.sap.iot.starterkit.ui/src/main/webapp/js/Component.js
@@ -1,56 +1,51 @@
-sap.ui.define( [ "sap/ui/core/UIComponent" ], function( UIComponent ) {
-	"use strict";
+sap.ui.core.UIComponent.extend( 'iotstarterkit.Component', {
 
-	return UIComponent.extend( "iotstarterkit.Component", {
-
-		metadata: {
-			routing: {
-				config: {
-					viewType: "JS",
-					viewPath: "js.view",
-					controlId: "iotstarterkitapp",
-					async: true
-				},
-				routes: [ {
-					pattern: "",
-					name: "main",
-					target: [ "main" ]
-				} ],
-				targets: {
-					main: {
-						viewName: "main",
-						viewId: "main",
-						viewLevel: 1,
-						controlAggregation: "pages"
-					}
+	metadata: {
+		routing: {
+			config: {
+				viewType: "JS",
+				viewPath: "js.view",
+				controlId: "iotstarterkitapp",
+				async: true
+			},
+			routes: [ {
+				pattern: "",
+				name: "main",
+				target: [ "main" ]
+			} ],
+			targets: {
+				main: {
+					viewName: "main",
+					viewId: "main",
+					viewLevel: 1,
+					controlAggregation: "pages"
 				}
 			}
-		},
-
-		init: function() {
-			UIComponent.prototype.init.apply( this, arguments );
-			this.initResourceModel();
-			this.getRouter().initialize();
-		},
-
-		initResourceModel: function() {
-			var sUrl = "locale/i18n.properties";
-			var oResourceModel = new sap.ui.model.resource.ResourceModel( {
-				bundleUrl: sUrl
-			} );
-			this.setModel( oResourceModel, "i18n" );
-		},
-
-		createContent: function() {
-			return sap.ui.view( {
-				viewName: "js.view.app",
-				type: "JS",
-				viewData: {
-					component: this
-				}
-			} );
 		}
+	},
 
-	} );
+	init: function() {
+		sap.ui.core.UIComponent.prototype.init.apply( this, arguments );
+		this.initResourceModel();
+		this.getRouter().initialize();
+	},
+
+	initResourceModel: function() {
+		var sUrl = "locale/i18n.properties";
+		var oResourceModel = new sap.ui.model.resource.ResourceModel( {
+			bundleUrl: sUrl
+		} );
+		this.setModel( oResourceModel, "i18n" );
+	},
+
+	createContent: function() {
+		return sap.ui.view( {
+			viewName: "js.view.app",
+			type: "JS",
+			viewData: {
+				component: this
+			}
+		} );
+	}
 
 } );
