+++ src/apps/java/consumption/com.sap.iot.starterkit.ui/src/main/webapp/js/view/app.view.js
@@ -1,9 +1,5 @@
 sap.ui.jsview( "js.view.app", {
 
-	getControllerName: function() {
-		return "js.controller.app";
-	},
-
 	createContent: function( oController ) {
 		this.setDisplayBlock( true );
 		return new sap.m.App( 'iotstarterkitapp' );
